+++ src/config/constants.py
@@ -4,7 +4,7 @@ Application Constants
 
 APP_NAME = "Book Library Reconciler"
 
-VERSION = "1.2.0"
+VERSION = "1.3.0"
 
 # --------------------------------------------------
 # Display